mainPage is now homePage / changes on route
