+++ src/conteiners/Router/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { ConnectedRouter } from "connected-react-router";
 import { Switch, Route } from "react-router-dom";
-import mainPage from '../MainPage/mainPage'
+import homepage from '../Home/homepage'
 
 export const routes = {
   root: "/",
@@ -11,7 +11,7 @@ function Router(props) {
   return (
     <ConnectedRouter history={props.history}>
       <Switch>
-        <Route exact path={routes.root} component={mainPage} />
+        <Route exact path={routes.root} component={homepage} />
       </Switch>
     </ConnectedRouter>
   );
